fixed wrong title being set in search dropdown (#135)
* fixed wrong title being set in search dropdown

* Refactored SearchBox component and defined Result, SearchData types

Relocated the types used by the SearchBox component ('Result' and 'SearchData') from inline in the component file to a separate types file. This simplifies the SearchBox component by removing lengthy type definitions. Alongside this, added a helper function 'getFirstSearchableFieldValue' which improves the logic for searchable fields within the component.

* [Refactoring] - Refine product validation in search box component

Introduced a helper function isValidProduct to validate search results based on productStatus, groupId, and categoryName in the search-box component. Additionally, updated how valid products are filtered and indexed to streamline and simplify the codebase.

* [chore] - Prettier Fix

* Add metadata to SearchBox component and expand SearchBox features

Added a metadata type for the SearchBox component in the storybook. Extended the features of the SearchBox by adding additional fields to the SearchData type and adjusting the SearchBox parameters to accommodate the new fields. Furthermore, discrimination of result type has been commented out for further review.

* fixed formatting

* fixed wrong title being set issue